Feat: add swagger Refactor: common packages
