+++ src/main/java/com/hscoreserver/hscorespring/common/config/HsConfig.java
@@ -1,4 +1,4 @@
-package com.hscoreserver.hscorespring.common;
+package com.hscoreserver.hscorespring.common.config;
 
 
 import lombok.Getter;
