+++ src/main/java/com/hscoreserver/hscorespring/common/model/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.hscoreserver.hscorespring.common;
+package com.hscoreserver.hscorespring.common.model;
 
 import java.time.LocalDateTime;
 import javax.persistence.Column;
