+++ src/main/resources/application-local.yml
@@ -1,4 +1,7 @@
 spring:
+  mvc:
+    pathmatch:
+      matching-strategy: ant_path_matcher
   datasource:
     driver-class-name: org.h2.Driver
     url: jdbc:h2:~/test
